[ELITERT-1221] Fix crashes with GitHub hosted repos
Fixes an issue that caused Dragnet to crash when generating an HTML
report for the results of an execution over a repository hosted on
Github.

The reason for the crash was that Ruby's URI parser is incapable of
parsing a GitHub URL, so, it was replaced by Git's URL.parse method,
which is capable of doing it.