Relaxes the version of the setup-ruby Github action
This is needed because the specific version that was being used suddenly
became unusable, because Github removed the ubuntu image linked to that
version.

To fix the issue the action is no longer fixed to a particular version,
instead it just takes v1 (the latest available within the same major).