[BUGFIX] Fix a crash when all repo verifiers pass
Fixes a crash that occurred in multi-repo mode when all the repos listed
in an MTR pass all the verifiers.

In this case the array of Repo objects was being returned by
ReposVerifier#verify. This return value was bubbling up all the way up
to Verifier#verify_mtr, which caused a NoMethodError to be raised
because the method was expecting a VerificationResult object, not the
array of Repo objects.