[ELITERT-1289] Handle Git errors during verification
Modifies the ChangesVerifier and the FilesVerifier class to handle the
case in which git is unable to determine the diff between two revisions.
In general this happens when an invalid or non-existing revision SHA1 is
added to a MTR.

Now, instead of Crashing dragnet will just report the MTR as failed.