[ELITERT-1204] Increase Ruby version to 3.1.0
Increase the minimum Ruby version for the project to 3.1.0. This is
needed in order to be able to update the 'git' gem to version 3.x.

This version of the 'git' gem includes a fix that prevents a crash when
calling the #branch method on a repository checked out in a detached
HEAD.